+++ src/components/AcceptInvitationForm.vue
@@ -88,12 +88,7 @@
         :disabled="btnDisabled"
         type="submit"
       >
-        <KIcon
-          v-if="currentState.matches('pending')"
-          color="currentColor"
-          icon="spinner"
-          :size="KUI_ICON_SIZE_30"
-        />
+        <ProgressIcon v-if="currentState.matches('pending')" class="spin-icon" :size="KUI_ICON_SIZE_40" />
         {{ btnText }}
       </KButton>
     </form>
@@ -101,7 +96,8 @@
 </template>
 
 <script setup lang="ts">
-import { inject, ref, Ref, reactive, computed, onMounted } from 'vue'
+import type { Ref } from 'vue'
+import { inject, ref, reactive, computed, onMounted } from 'vue'
 import { createMachine } from 'xstate'
 import { useMachine } from '@xstate/vue'
 import { win } from '@/utils'
@@ -109,10 +105,11 @@ import useConfigOptions from '@/composables/useConfigOptions'
 import useI18n from '@/composables/useI18n'
 import useAxios from '@/composables/useAxios'
 import { acceptInvitationEmits } from '@/components/emits'
-import { AxiosResponse } from 'axios'
-import { KUI_ICON_SIZE_30 } from '@kong/design-tokens'
+import type { AxiosResponse } from 'axios'
+import { ProgressIcon } from '@kong/icons'
+import { KUI_ICON_SIZE_40 } from '@kong/design-tokens'
 // Components
-import { KButton, KIcon, KInput } from '@kong/kongponents'
+import { KButton, KInput } from '@kong/kongponents'
 import ErrorMessage from '@/components/ErrorMessage.vue'
 
 const emit = defineEmits(acceptInvitationEmits)